requirements.txt
numpy
pillow
tensorflow
scikit-learn
matplotlib
seaborn
scikit-plot
jupyterlab
fastapi
uvicorn
streamlit
requests
python-multipart
pymongo
python-dotenv